LPS-77631 Move adaptive.media.internal package to adaptive-media-web
