+++ modules/apps/adaptive-media/adaptive-media-web/src/main/java/com/liferay/adaptive/media/web/internal/messaging/AMMessagingConfigurator.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.adaptive.media.internal.messaging;
+package com.liferay.adaptive.media.web.internal.messaging;
 
-import com.liferay.adaptive.media.internal.configuration.AMConfiguration;
-import com.liferay.adaptive.media.internal.constants.AMDestinationNames;
+import com.liferay.adaptive.media.web.internal.configuration.AMConfiguration;
+import com.liferay.adaptive.media.web.internal.constants.AMDestinationNames;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.messaging.Destination;
 import com.liferay.portal.kernel.messaging.DestinationConfiguration;
@@ -37,7 +37,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Adolfo Pérez
  */
 @Component(
-	configurationPid = "com.liferay.adaptive.media.internal.configuration.AMConfiguration",
+	configurationPid = "com.liferay.adaptive.media.web.internal.configuration.AMConfiguration",
 	immediate = true
 )
 public class AMMessagingConfigurator {
